HDX-9753 last_modified change on url change only for external resources
for uploaded resources, it can happen that it looks like the url has changes although the file didn't change:
https://data.humdata.local/dataset/DATASET_ID/resource/RESOURCE_ID/download/test1.csv VS test1.csv

+ tests
